+++ netbox/dcim/forms.py
@@ -13,7 +13,7 @@ from timezone_field import TimeZoneFormField
 from extras.forms import AddRemoveTagsForm, CustomFieldForm, CustomFieldBulkEditForm, CustomFieldFilterForm
 from ipam.models import IPAddress, VLAN, VLANGroup
 from tenancy.forms import TenancyForm
-from tenancy.formset import TenancyFilterForm
+from tenancy.forms import TenancyFilterForm
 from tenancy.models import Tenant, TenantGroup
 from utilities.forms import (
     APISelect, APISelectMultiple, add_blank_choice, ArrayFieldSelectMultiple, BootstrapMixin, BulkEditForm,
